* 'windows-fixes' of https://github.com/Tarjei400/TibiaEditor:
  Support 9.6
  Added 9.6 support

+++ tibiasprite.h
@@ -1,8 +1,8 @@
 #ifndef TIBIASPRITE_H
 #define TIBIASPRITE_H
 
-#include <QImage>
-#include <QHash>
+#include <QtGui/QImage>
+#include <QtCore/QHash>
 
 #include "tibiaobject.h"
 
@@ -15,7 +15,7 @@
 class TibiaSprite : public TibiaObject
 {
 public:
-    TibiaSprite( quint16 _id = 0, quint8 _width = 1, quint8 _height = 1 );
+    TibiaSprite( quint32 _id = 0, quint8 _width = 1, quint8 _height = 1 );
     virtual ~TibiaSprite( void );
 
     bool isDummy() const {
@@ -27,7 +27,7 @@ public:
     void setImage( const QImage& );
     static void formatImage( QImage& image );
 
-    quint16 id;
+    quint32 id;
     quint8 width, height, r, g, b;
     QImage image;
     bool dummy;
